fix monitoring metric aggregations on cluster dash
