+++ models/management.cattle.io.cluster.js
@@ -1,5 +1,5 @@
 import { CATALOG } from '@/config/labels-annotations';
-import { FLEET, MANAGEMENT, NODE } from '@/config/types';
+import { FLEET, MANAGEMENT } from '@/config/types';
 import { insertAt } from '@/utils/array';
 import { downloadFile } from '@/utils/download';
 import { parseSi } from '@/utils/units';
@@ -313,36 +313,6 @@ export default {
       const out = jsyaml.dump(obj);
 
       downloadFile('kubeconfig.yaml', out, 'application/yaml');
-    };
-  },
-
-  fetchNodeMetrics() {
-    return async() => {
-      const nodes = await this.$dispatch('cluster/findAll', { type: NODE }, { root: true });
-      const nodeMetrics = await this.$dispatch('cluster/findAll', { type: NODE }, { root: true });
-
-      const someNonWorkerRoles = nodes.some(node => node.hasARole && !node.isWorker);
-
-      const metrics = nodeMetrics.filter((metric) => {
-        const node = nodes.find(nd => nd.id === metric.id);
-
-        return node && (!someNonWorkerRoles || node.isWorker);
-      });
-      const initialAggregation = {
-        cpu:    0,
-        memory: 0
-      };
-
-      if (isEmpty(metrics)) {
-        return null;
-      }
-
-      return metrics.reduce((agg, metric) => {
-        agg.cpu += parseSi(metric?.usage?.cpu);
-        agg.memory += parseSi(metric?.usage?.memory);
-
-        return agg;
-      }, initialAggregation);
     };
   },
